Use target all for build instead of skunit_tests
This fixes an issue where resources for tests (such as images) aren't copied,
causing tests to fail. This happens because building a specific executable target
such as skunit_tests doesn't run the custom targets used for the copy. This
could be fixed by altering CMakeLists.txt but updating this guide is simpler.